VS-90 Make illustration query
